Player: extract handlers
