+++ src/components/Player/Player.jsx
@@ -13,6 +13,8 @@ import GridBuffer from './GridBuffer'
 import InputBox from './InputBox'
 import Status from './Status'
 
+import {Handlers} from './playerHandlers'
+
 import './player.css'
 
 const INITIAL_STATUS = {
@@ -29,53 +31,6 @@ const runMachine = ({ engine: Engine, file, handlers }) => {
 
   return { sendFn, instance: vm }
 }
-
-const Handlers = ({
-  setStatus,
-  setWindows,
-  setCurrentWindowId,
-  setInputType,
-  setInbox
-}) => ({
-  onInit: _ => {
-    setStatus({ stage: 'ready' })
-  },
-  /* */
-  onUpdateWindows: windows => {
-    setWindows(windows)
-  },
-  onUpdateInputs: data => {
-    if (data.length === 0) return void null
-
-    const {type, id} = data[0]
-    setCurrentWindowId(id)
-    setInputType(type)
-  },
-  onUpdateContent: inbox => {
-    setInbox(inbox)
-  },
-  onDisable: _ => {
-    setInputType(null)
-  },
-  /* */
-  onFileNameRequest: (tosave, usage, _, setFileName) => {
-    setFileName({
-      usage,
-      filename: prompt('Enter the filename')
-    })
-  },
-  onFileRead: ({ filename }) => {
-    const content = localStorage.getItem(`fake-fs/${filename}`)
-    return decode(content)
-  },
-  onFileWrite: ({ filename }, content) => {
-    localStorage.setItem(`fake-fs/${filename}`, encode(content))
-  },
-  /* */
-  onExit: _ => {
-    setInputType(null)
-  }
-})
 
 export default function ({
   vmParts: { file, engine }, singleWindow
